# docs/swagger.yaml
components:
  schemas:
...
    BaseResponse:
      properties:
        msg:
          default: ''
          title: Msg
          type: string
        status:
          default: success
          title: Status
          type: string
      title: BaseResponse
      type: object
    BlockDataResponse:
      properties:
        content:
          anyOf:
          - type: string
          - $ref: '#/components/schemas/WikiLinkSchema'
          title: Content
        created_at:
          format: date-time
          title: Created At
          type: string
        document_id:
          format: uuid
          title: Document Id
          type: string
        id:
          format: uuid
          title: Id
          type: string
        link:
          anyOf:
          - type: string
          - type: 'null'
          title: Link
        permission_mode:
          $ref: '#/components/schemas/ObjectPermissionMode'
        position:
          title: Position
          type: integer
        type_block:
          $ref: '#/components/schemas/TypeBlock'
      required:
      - permission_mode
      - id
      - document_id
      - position
      - type_block
      - created_at
      - content
      title: BlockDataResponse
      type: object
...
    Body_update_block_data_file_api_v1_blocks_data_file_put:
      properties:
        file:
          format: binary
          title: File
          type: string
      required:
      - file
      title: Body_update_block_data_file_api_v1_blocks_data_file_put
      type: object
    CreateBlock:
      properties:
        document_id:
          format: uuid
          title: Document Id
          type: string
        position:
          title: Position
          type: integer
        type_block:
          $ref: '#/components/schemas/TypeBlock'
      required:
      - document_id
      - position
      - type_block
      title: CreateBlock
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    UpdateBlockData:
      properties:
        block_id:
          format: uuid
          title: Block Id
          type: string
        content:
          title: Content
          type: string
      required:
      - block_id
      - content
      title: UpdateBlockData
      type: object
    UpdateBlockInfo:
      properties:
        block_id:
          format: uuid
          title: Block Id
          type: string
        position:
          anyOf:
          - type: integer
          - type: 'null'
          title: Position
        type_block:
          anyOf:
          - $ref: '#/components/schemas/TypeBlock'
          - type: 'null'
      required:
      - block_id
      title: UpdateBlockInfo
      type: object
...
    VersionBlockInfo:
      properties:
        commit_id:
          maxLength: 64
          minLength: 64
          title: Commit Id
          type: string
        committer_user:
          $ref: '#/components/schemas/UserBaseInfoResponse'
        created_at:
          format: date-time
          title: Created At
          type: string
        object_id:
          format: uuid
          title: Object Id
          type: string
      required:
      - commit_id
      - object_id
      - committer_user
      - created_at
      title: VersionBlockInfo
      type: object
...
    WikiLinkSchema:
      properties:
        block_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Block Id
        document_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Document Id
        workspace_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Workspace Id
      title: WikiLinkSchema
      type: object
...
  securitySchemes:
    WikiAccessToken:
      in: cookie
      name: wiki_access_token
      type: apiKey
    WikiApiKey:
      in: header
      name: X-Wiki-API-Key
      type: apiKey
    WikiBearer:
      scheme: bearer
      type: http
    WikiRootApiKey:
      description: 'Root key, allows you to perform actions as administrator, random
        user data is generated.

        **Limitation:** operations requiring an explicitly registered userare unavailable,
        example.: `Create: POST /api_key/`.'
      in: header
      name: X-Wiki-ROOT-API-Key
      type: apiKey
info:
  description: Welcome to Wiki's API documentation! Here you will able to discover
    all of the ways you can interact with the Wiki API.
  title: Wiki API
  version: 0.0.1
openapi: 3.1.0
paths:
...
  /api/v1/blocks/:
    delete:
      operationId: delete_block_api_v1_blocks__delete
      parameters:
      - in: query
        name: block_id
        required: true
        schema:
          format: uuid
          title: Block Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - WikiRootApiKey: []
      - WikiApiKey: []
      - WikiApiKey: []
      - WikiAccessToken: []
      - WikiBearer: []
      summary: Delete a block from a document
      tags:
      - WorkspaceBlocks
    post:
      operationId: create_block_api_v1_blocks__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBlock'
        required: true
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockDataResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - WikiRootApiKey: []
      - WikiApiKey: []
      - WikiApiKey: []
      - WikiAccessToken: []
      - WikiBearer: []
      summary: Create block
      tags:
      - WorkspaceBlocks
  /api/v1/blocks/data:
    get:
      operationId: get_block_by_id_api_v1_blocks_data_get
      parameters:
      - in: query
        name: block_id
        required: true
        schema:
          format: uuid
          title: Block Id
          type: string
      - in: query
        name: version_commit_id
        required: false
        schema:
          anyOf:
          - maxLength: 64
            minLength: 64
            type: string
          - type: 'null'
          title: Version Commit Id
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockDataResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - WikiRootApiKey: []
      - WikiApiKey: []
      - WikiApiKey: []
      - WikiAccessToken: []
      - WikiBearer: []
      summary: Get data block by id
      tags:
      - WorkspaceBlocks
  /api/v1/blocks/data/all:
    get:
      description: '## Getting document blocks data


        :param document_id: Document Identifier

        :param version_commit_id: Specify the commit identifier of a particular version.
        If None, get the latest version of the data.'
      operationId: get_blocks_api_v1_blocks_data_all_get
      parameters:
      - in: query
        name: document_id
        required: true
        schema:
          format: uuid
          title: Document Id
          type: string
      - in: query
        name: version_commit_id
        required: false
        schema:
          anyOf:
          - maxLength: 64
            minLength: 64
            type: string
          - type: 'null'
          title: Version Commit Id
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/BlockDataResponse'
                title: Response Get Blocks Api V1 Blocks Data All Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - WikiRootApiKey: []
      - WikiApiKey: []
      - WikiApiKey: []
      - WikiAccessToken: []
      - WikiBearer: []
      summary: Get all data blocks by id document
      tags:
      - WorkspaceBlocks
  /api/v1/blocks/data/file:
    put:
      operationId: update_block_data_file_api_v1_blocks_data_file_put
      parameters:
      - in: query
        name: block_id
        required: true
        schema:
          format: uuid
          title: Block Id
          type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_update_block_data_file_api_v1_blocks_data_file_put'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockDataResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - WikiRootApiKey: []
      - WikiApiKey: []
      - WikiApiKey: []
      - WikiAccessToken: []
      - WikiBearer: []
      summary: Update file data document block
      tags:
      - WorkspaceBlocks
  /api/v1/blocks/data/link:
    put:
      description: '## Update wiki link data document block

        One of the specified object ids will be used to create the link in priority:
        block_id, document_id, workspace_id'
      operationId: update_block_data_link_api_v1_blocks_data_link_put
      parameters:
      - in: query
        name: block_id
        required: true
        schema:
          format: uuid
          title: Block Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WikiLinkSchema'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockDataResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - WikiRootApiKey: []
      - WikiApiKey: []
      - WikiApiKey: []
      - WikiAccessToken: []
      - WikiBearer: []
      summary: Update wiki link data document block
      tags:
      - WorkspaceBlocks
  /api/v1/blocks/data/text:
    put:
      operationId: update_block_data_text_api_v1_blocks_data_text_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBlockData'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockDataResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - WikiRootApiKey: []
      - WikiApiKey: []
      - WikiApiKey: []
      - WikiAccessToken: []
      - WikiBearer: []
      summary: Update text data document block
      tags:
      - WorkspaceBlocks
  /api/v1/blocks/info:
    put:
      operationId: update_block_info_api_v1_blocks_info_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBlockInfo'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockInfoResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - WikiRootApiKey: []
      - WikiApiKey: []
      - WikiApiKey: []
      - WikiAccessToken: []
      - WikiBearer: []
      summary: Update info document block
      tags:
      - WorkspaceBlocks
  /api/v1/blocks/publish/data:
    get:
      description: '## Getting published document blocks data'
      operationId: get_publish_blocks_api_v1_blocks_publish_data_get
      parameters:
      - in: query
        name: document_id
        required: true
        schema:
          format: uuid
          title: Document Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/BlockDataResponse'
                title: Response Get Publish Blocks Api V1 Blocks Publish Data Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - WikiRootApiKey: []
      - WikiApiKey: []
      - WikiApiKey: []
      - WikiAccessToken: []
      - WikiBearer: []
      summary: Get all publish data blocks by id document
      tags:
      - WorkspaceBlocks
...
  /api/v1/versioning/block/{block_id}/info:
    get:
      operationId: get_list_versions_document_block_api_v1_versioning_block__block_id__info_get
      parameters:
      - in: path
        name: block_id
        required: true
        schema:
          format: uuid
          title: Block Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/VersionBlockInfo'
                title: Response Get List Versions Document Block Api V1 Versioning
                  Block  Block Id  Info Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - WikiRootApiKey: []
      - WikiApiKey: []
      - WikiApiKey: []
      - WikiAccessToken: []
      - WikiBearer: []
      summary: Get info as a list of all versions block
      tags:
      - WorkspaceVersioning
...
servers:
- description: Local server
  url: http://localhost:8000/
- description: Demo server
  url: https://api.wiki.example.com/
- description: Ngrok free server
  url: https://poorly-ideal-cobra.ngrok-free.app/